Added backlight control to setup tile
