+++ src/gui/mainbar/setup_tile/setup_tile.h
@@ -25,21 +25,24 @@
     #include <TTGO.h>
     #include "gui/icon.h"
 
-    #define MAX_SETUP_ICON_HORZ       3
-    #define MAX_SETUP_ICON_VERT       3
+    #define MAX_SETUP_ICON_HORZ       2
+    #define MAX_SETUP_ICON_VERT       1
     #define MAX_SETUP_TILES           1
     #define MAX_SETUP_ICON            ( MAX_SETUP_ICON_HORZ * MAX_SETUP_ICON_VERT * MAX_SETUP_TILES )
 
     #define SETUP_ICON_X_SIZE         64
     #define SETUP_ICON_Y_SIZE         64
-    #define SETUP_ICON_X_CLEARENCE    8
-    #define SETUP_ICON_Y_CLEARENCE    8
+    #define SETUP_ICON_X_CLEARENCE    64
+    #define SETUP_ICON_Y_CLEARENCE    30
     //#define SETUP_ICON_Y_CLEARENCE    36
     #define SETUP_LABEL_X_SIZE        SETUP_ICON_X_SIZE + SETUP_ICON_X_CLEARENCE
     #define SETUP_LABEL_Y_SIZE        SETUP_ICON_Y_CLEARENCE / 2
 
+    //#define SETUP_FIRST_X_POS         ( 240 - ( SETUP_ICON_X_SIZE * MAX_SETUP_ICON_HORZ + SETUP_ICON_X_CLEARENCE * ( MAX_SETUP_ICON_HORZ - 1 ) ) ) / 2
     #define SETUP_FIRST_X_POS         ( 240 - ( SETUP_ICON_X_SIZE * MAX_SETUP_ICON_HORZ + SETUP_ICON_X_CLEARENCE * ( MAX_SETUP_ICON_HORZ - 1 ) ) ) / 2
-    #define SETUP_FIRST_Y_POS         ( 240 - ( SETUP_ICON_Y_SIZE * MAX_SETUP_ICON_VERT + SETUP_ICON_Y_CLEARENCE * ( MAX_SETUP_ICON_VERT - 1 ) ) ) / 2
+    //#define SETUP_FIRST_Y_POS         ( 240 - ( SETUP_ICON_Y_SIZE * MAX_SETUP_ICON_VERT + SETUP_ICON_Y_CLEARENCE * ( MAX_SETUP_ICON_VERT - 1 ) ) ) / 2
+    //fill icons from bottom
+    #define SETUP_FIRST_Y_POS         ( 240 - ( SETUP_ICON_Y_SIZE * MAX_SETUP_ICON_VERT + SETUP_ICON_Y_CLEARENCE ) )
 
     /**
      * @brief setup the setup tile
