+++ src/gui/mainbar/setup_tile/wlan_settings/wlan_settings.h
@@ -24,9 +24,11 @@
 
     #include <TTGO.h>
 
-    void wlan_settings_tile_setup( void );
+    void wlan_connection_tile_pre_setup( void );
+    void wlan_connection_tile_setup( void );
     void wlan_password_tile_setup( uint32_t wifi_password_tile_num );
-    void wlan_setup_tile_setup( uint32_t wifi_setup_tile_num );
+    void wlan_setup_tile_pre_setup( void );
+    void wlan_setup_tile_setup( void );
     void wlan_setup_display_ssid(String ssid);
 
 #endif // _WLAN_SETTINGS_H
