+++ src/gui/setup.h
@@ -57,5 +57,9 @@
 
     void setup_add_submenu_tile(void);
 
+    uint32_t setup_get_submenu2_tile_num();
+
+    void setup_add_submenu2_tile(void);
+
 
 #endif // _SETUP_H
